validare date Medicament

+++ src/main/java/com/clinica/web/dto/MedicamentDto.java
@@ -15,7 +15,7 @@ public class MedicamentDto {
     private int medicamentID;
     private String nume;
     private int stoc;
-    private int pret;
+    private float pret;
     private LocalDate data_expirarii;
 
 }
